refactor products api to separate implementation
